fix the plugin file
