--- conftest.py
@@ -1,4 +1,0 @@
-pytest_plugins = [
-    "pytest_helm_charts.giantswarm_app_platform.apps.http_testing",
-    "pytest_helm_charts.giantswarm_app_platform.fixtures",
-]
