+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pytest-helm-charts"
-version = "0.1.2"
+version = "0.1.3"
 description = "A plugin to provide different types and configs of Kubernetes clusters that can be used for testing."
 authors = ["Łukasz Piątkowski <lukasz@giantswarm.io>"]
 license="Apache-2.0"
@@ -28,7 +28,7 @@ classifiers=[
 [tool.poetry.plugins] # Optional super table
 
 [tool.poetry.plugins."pytest11"]
-"helm-charts" = "pytest_helm_charts.fixtures"
+"helm-charts" = "pytest_helm_charts.plugin"
 
 [tool.poetry.dependencies]
 python = "^3.7"
